Unificando clases y detalles
Unificando clases para no tener conocimientos de mas
