+++ src/morfiYA/domain/Cliente.java
@@ -60,11 +60,4 @@ public class Cliente {
 		    creditos += cantidad;
 		 
 	}
-	public void setCreditos(double creditos) {
-		this.creditos = creditos;
-	}	
-	
-	public void update(double monto) {
-		setCreditos(creditos + monto);
-	}
 }
